Move a shared routine to the shared class
